requirements.txt
fastapi==0.116.1
openai==1.96.1
pandas==2.2.3
pdfplumber==0.11.7
pinecone-client>=4.0.0
pydantic==2.11.7
python-dotenv==1.0.1
uvicorn[standard]
python-multipart
gradio==4.44.0

